Update Docker base images to bookworm
We've updated many other things from Debian bullseye to Debian bookworm, so it
seems reasonable to also update the EDD base images just to keep things
consistent. Note, even though this is using the slim version of the base image,
the overall weight is still increasing by about 48MB, and about 35MB of that is
entirely accounted for by the base image (base image 121MB for bullseye, 156MB
for bookworm; final EDD 713MB for bullseye, 761MB for bookworm). The sizes are
still well within tolerances, but we should be aware of the larger size and
look for further opportunities to slim down the images when possible.