Remove dead code
Our custom handling code for dealing with signups and LDAP accounts moved to
the `edd` module more than three years ago. The previous location was kept as a
shim, reporting a warning if used. Three years ought to be enough for anyone to
notice the warning and migrate to the new name for settings values.