Add tests to `edd.account` module
The code handling our custom account handling was one of the big remaining
outliers on code coverage. Adding a bunch of tests to these, and cleaning up
some of the code that was not operating as expected.

Coverage is increased from 29% to 76%, with the remaining uncovered portions
being code that is going to be much more difficult to cover, as it involves
accounts logging in from LDAP or from social account identity providers. These
will necessarily involve either extensive setup, or mocking, in order to
properly test. While these would be useful to have, it's a lot of effort for
minimal gain. The remaining coverage is a "maybe someday" priority.