Switch to `bun install` from `yarn install`
The `bun` JavaScript runtime, bundler, installer, etc. recently released its
1.0 version, and one of its main selling points is its `npm`-compatible
installer will create a `node_modules` folder for an app much faster than
`npm`, `pnpm`, or `yarn` will do so. Switching over results in the build phase
for grabbing these dependencies running about 125% faster (from about 45sec to
about 20sec).

The bundler and runtime are not quite out-of-the-box replacements yet, so only
switching the package manager portion of the build with `bun` at the moment.
We're still using the official Node Alpine image as a base, and simply copying
the `node_modules` folder from the `oven/bun` image for each of the
pre-Bootstrap-5 and post-Bootstrap-5 frontend build environments.

There are some issues with getting the Jest unit tests to run, as the install
of `canvas` does not properly complete in some environments. To fix in the
short term, removing the JS unit tests. These were already very low-signal,
only covering a fraction of one percent of the non-Bootstrap-5 TypeScript.
While it would be nice to have useful unit tests, at the moment this is not
useful and should not be enabled.