Fix parse error in Experiment Description code
In some cases, an empty row in an Excel file is processed as if it has content,
but it doesn't. This causes EDD to hit an IndexError when addressing a cell in
the empty row that does not "exist". Adding some guard checks to abort
processing when the row does not have any cells at the given index.