Improve handling of transcriptomics type lookups
Handling types for transcriptomics is a bit different than all the other types,
as the EDD stakeholders have not settled on any "blessed" external source(s)
for names. The app supports looking up entries in the linked ICE instance as
the only supported "global" names. All others are treated as user-scoped names.
When looking up a gene name, e.g. "X-23", if the linked ICE has a part "X-23",
then that is the type. If Alice adds data for "X-23", and it's *not* in the
linked ICE, then EDD creates an "X-23" type unique to Alice. If Barbara also
adds data for "X-23", then EDD creates an "X-23" distinct from the one created
for Alice.

This behavior is also our work-around for the difficulty of adding new types to
EDD. Where the "official" way to add a type is to use the identifier from the
"blessed" external database, like PubChem; or, request an admin add the type
for you. Instead, importing data as "omics" would fall back to treating
anything unknown as a `GeneIdentifier`.

Eventually, we want all types in EDD to be mapped to canonical references in
external databases, i.e. NCBI for genes. There is a flag `provisional` on every
type record, which is meant to indicate that it's a type created in EDD without
validation from one of these external databases. However, the `GeneIdentifier`
models were not making use of the flag.

This update changes the code responsible for looking up and creating the
`GeneIdentifier` model records, to make it aware of the `provisional` flag. An
out-of-band database update will execute to bring existing records in line with
how new records will be created. The code for this update is:

```python
from main.models import GeneIdentifier
GeneIdentifier.objects.filter(strainlink__isnull=True).update(provisional=True)
```

The result of the change here will not be visible to ordinary users of EDD, nor
to web-interface-only admins of EDD. It will enable the updated import code to
better make a small change in behavior, along with its new features. As the new
import code will allow for choosing a measurement type from an autocomplete
search if none is found, there will be less need to rely on EDD to correctly
guess the type intended by users, and less need to automatically create new
types. The plan is to have the updated import not automatically create new
types during the lookup phase, and instead push the type creation to the
"Interpret" phase. This will result in fewer "bogus" types being created in the
database from failed, partial, or erroneous import files.