Move threadlocal conftest fixture to root
An auto-use fixture used to clear out anything in the `request` threadlocal
variable is moved to the project Python root, instead of in the `main` module.
When the fixture is in the `main` module, it only applies to tests in that
module, while EDD has tests in multiple modules that could manipulate the
`request` threadlocal value.

This helps make tests more isolated and consistent in modules outside of `main`.