Improve autocomplete searches
Adding HTML versions of items to results of autocomplete searches. Add support
for an autocomplete to include a first item for creating a new record from that
autocomplete. This will require the form to know about the creation option.

Dropping refactored `optional_sort` method on autocompletes; it's not yet used,
and is incompatible with some of the union-type SQL queries and with any
Solr-based queries.