Remove AJAX-specific versions of URLs
Many of the inline-update portions of the refreshed Bootstrap 5 interface made
use of specific URLs for views to render partial pages. These are removed, in
favor of checking the "X-Requested-With" header that JQuery will automatically
include in AJAX requests. This way, both full page refreshes and partial
updates can use the same URLs, and the view can choose which method is being
used via the header value.