Add privlege drop to run_tests.sh script
EDD's normal execution uses an entrypoint script that runs as `root` for
initial setup, then `exec`s the application itself as the `edduser` account.
This can cause a clash when developing EDD, and using both the application and
running tests with the builtin `run_tests.sh` script as `root`. The test runner
would write directories to disk as `root`, and the application could try to
also write to directories owned by `root`, when the expectation is the
directory would be owned by `edduser`.

This update adds a check to `run_tests.sh`, and when it's executed by `root`,
the script will `exec` itself with `setpriv` to drop execution privleges to
`edduser` similar to how the entrypoint script drops to `edduser` with `gosu`.
The mechanism here is a little different, using `setpriv` instead of `gosu`;
the latter is specifically designed for dropping privleges in a Docker
entrypoint, and has potential issues when used from an interactive TTY (as
when `docker exec` into the container to run tests interactively).