Refactor utility code for dropzones
Code for dropzones has gone through a lot of changes in the years of EDD's
development. The interface refresh to Bootstrap 5 initially brought over a lot
of legacy from the prior use-cases to the structure of the code. These other
use-cases either do not exist in the Bootstrap 5 version of the interface, or
have not yet migrated. As the usage is simpler, we can simplify the interfaces.

Dropping the custom `DZOptions` interface, which was meant to bridge options
from different dropzone libraries. Instead, using only `dropzone.js`, and using
its API options directly. Also dropping the `createDropzone` function, now that
code is directly using the `Dropzone` constructor.