Update base style customizations sizing
Our base styles make use of `em` units in a few places, where they need to
instead use `rem`. The former are relative units to the parent element, and
thus suceptible to compunding; while the latter are relative to the document
root. See: https://www.digitalocean.com/community/tutorials/css-rem-vs-em-units

Only a few items in our base style were using `em` when `rem` is more
appropriate. Adding in a fix rule for our use of FontAwesome; the default rules
for a "stack" of icons uses `em` units, which makes the "stack" look unexpected
when parent elements are already sized. Forcing the sizing of the FontAwesome
class to use `rem` makes the icon stack appear normally.