Add support for keyword arguments to study_reverse
The Study view mixins offer a `study_reverse` method, which will allow for
getting a URL for a Study view, using the same study as the current view. This
is fine for simple views that only have URL arguments for the Study, but
doesn't work when there's additional parameters to the generated URL. Adding a
keyword argument double splat allows for any keywords *other* *than* the Study
argument to be passed, while still auto-filling the Study argument.

All current uses of the method will continue to work as-is. Any URL specs that
use arguments in addition to the Study ID / UUID / slug can now use them by
passing directly; e.g.:

    self.study_reverse("main:study_sprox", page=42)

This has no uses in existing code yet, but will be required for some upcoming
features. Stay tuned ...