Add overhauled import workflow
We have a need to update several parts of the data ingestion solutions
available in EDD, and in particular adding support for setting metadata with
data, during one import file upload. There is also a need to boost performance,
particularly around bulk database inserts.

The existing APIs for these features are too rigid to allow for comfortable
modifications. The `reader` module is an attempt to refresh the APIs, to do the
same functionality while allowing for the new features to be added as well. The
module contains replacements for much of the code in the
`edd.load.parsers.core` module. A main focus is to simplify the APIs and to
support iteration and incremental updates as much as possible.

All frontend code is replaced. Where the existing frontend is the only piece of
EDD making use of React, and works as a single-page app embedded in EDD and
interacts with custom REST APIs, the new frontend uses traditional HTML forms
and minimal JavaScript. The little JavaScript it includes is mainly to
facilitate inline updates of progress bars--the primary "new" feature of the
frontend. Otherwise, the layout and functionality is purposefully mimicing the
existing frontend, with minor updates for a11y and to switch to Bootstrap 5.

This is a __*LOT*__ of code removed. All the React front-end is gone. Most of
the custom and flaky error reporting framework is ripped out; the custom
Exception attributes are kept in place, but the rest is removed. We can adapt
these later to more standard error reporting. A lot of the database models are
no longer needed. The REST API subsystem for imports is gone. All the resolving
code that relied on loading the entire import in memory is gone, in favor of
the newer generator/iterator based resolving. The clone of the notification
system specific to the React front-end is gone, in favor of the simpler
progress bar update websocket.