Update legacy strain page to list Studies
Having a single link for every legacy Strain isn't as useful as having a
distinct link, which also lists the studies referencing the strain. This gets
us at least to the level we would have for a blank ICE entry for the strain,
without actually creating more garbage in ICE to do the migration.