Update Python base to 3.12
Along with the Python update, locking dependencies as well. This includes a
Django update, and a few settings APIs are changing with new Django. The
Allauth app also has a few changes requiring settings updates and a minor
refactoring of names in our custom account handling code. Pytest has changed
how it handles custom test markers, our custom marker for ICE integration code
is now in the `pyproject.toml` settings for pytest to silence the warnings.