Fix line editing when a strain value is present
Jennifer reported a bug where she was unable to edit line names in EDD. After
further investigation, I found that the bug was not with editing line names,
but doing any edits on lines containing strain links.

There are a few interacting bugs at play. One that delayed diagnosis of the
root cause is a bug where the strain autocomplete from ICE was sending EDD the
ICE database primary key value, where it should have sent the strain UUID. This
is updated now to return the correct values. There remains a TODO for a future
enhancement, to make use of the additional data ICE makes available on search
results, so EDD can display more than just the entry name in ICE.

The true root cause of the errors hit by Jennifer have to do with the order of
operations Django is using to validate values to a ModelChoiceField. We have a
RegistryValidator class that will auto-create EDD Strain model entries from an
ICE entry; but this only runs after the Django ModelChoiceField does its
validations. So we're stuck in a spot where we can't validate a UUID matches
an ICE strain until we have a saved model instance, and we can't have a saved
model instance until we validate a UUID matches an ICE strain.

After some experimentation, I settled on making a custom Field class, which
extends ModelMultipleChoiceField, and ensure that the translation of values
from a form submit to the values validated by the Field first run through
RegistryValidator. This fixes the order of operations, so we check ICE first,
creating an EDD model instance if needed, then do foreign key database ops.

Fixes: #1308