Add websocket update on every import transition
When an import transitions, especially to the FAILED state, there was a
possiblity of the user not getting a notification for the transition. This
would appear as the import progress bar stalling. Reloading the page would
refresh the websocket, and fetch the failure to turn the progress bar red. This
tweaks the code so all transitions trigger a message to any websocket
listeners, and replaces the few locations where an explicit update call is made
immediately after a transition call.