Fix handling of export button submit events
A few interactions within the description page caused exports and the worklist
generation feature in particular to not work. The buttons for export and the
menu of items in the combo button have a click handler. This handler is
supposed to check the state of the line table selection, and pass the
appropriate IDs to the export page POST request. However, there is also a
handler for all forms on the page, meant to do validation of inline forms and
do inline replacements. The submit event on the click handler gets intercepted,
and the chosen export page ends up not appearing.

Updating the page layout a bit, so that:
1) the export buttons get their own form, which is excluded from the other
   form handling; and,
2) the selected items from the lines table get copied to the export form prior
   to submitting;

Fixes: #1312