Fix Typescript build in Docker
Something has shifted in the Javascript dependencies for both versions of EDD's
front-end build, and now we must explicitly add a dependency on `ajv` to
correctly resolve all the packages.

The error would look something like:

    Error: Cannot find module 'ajv/dist/compile/codegen'

We aren't directly using the package, but other things are. Presumably there's
an incompatibility in the version that one thing is installing and another, but
I don't have the bandwidth to diagnose further, and pinning it to v8 seems to
fix the problem.