Add poetry-export-plugin
Using the `poetry export` sub-command generates warnings about the sub-command
being removed from Poetry in the future, and moving instead to a plugin.
Installing that plugin, so we can continue using it to create a plain
`requirements.txt` to do the final build Python environment creation. Also
adding a config flag to disable the warning in Poetry, as installing the plugin
does *not* silence the warning.

Fixes: #1316