Fix Description view for user with Read access
When viewing a Study where the logged in user only has Read level access, the
lines table in the Description view would be empty. This turned out to be
caused by the drag-n-drop element to add more Line entries with an Experiment
Setup File, which would not be added to the page for a user who could not write
new Lines. The setup code for the table would still try to setup the dropzone
element, but fail when trying to index the first entry in the selector for the
dropzone. This failure then causes the rest of the table setup to not execute.
Updating the dropzone module to first check if the selector passed to the
initialization code has one (and only one) item, allows for the code to safely
skip over the missing element, and continue rendering the table.

Fixes: #1314