Remove deprecation check on metadata
The default metadata used by EDD itself had a change in API a while ago,
replacing a constant with a lookup function. There was a deprecation check on
using the old constant for a while, with a comment to remove it when 2.8.0 is
reached. Since we've updated the version, the checks are now removed.