Remove Phosphor measurement type
The commonly-used measurement types in EDD right now are:
- "Generic", measurements without any specific external references or other
  metadata, e.g. "OD600", "pH"
- "Metabolite", measurements of small molecules, almost all of which have a
  reference to PubChem
- "ProteinIdentifier", measurements of peptides, almost all of which have a
  reference to UniProt or ICE
- "GeneIdentifier", measurements of RNA transcripts; these mostly *don't* have
  external references, though the plan is to support NCBI and/or ICE

There's also one other measurement type, "Phosphor", which was meant to be a
measurement of specific light wavelengths from phosphorescence / fluorescence /
luminescence. It was only ever used once, for the Reider Apel et al paper using
the flow cytometer to characterize promoter expression levels. The idea was we
would eventually get the analysis pipeline from Jake Beal to better quantize
signals from the flow cytometer, and it would be good to have measurement types
that referenced the analysis, had metadata on excitation and emission
wavelengths, etc. But, this never happened. And we have a bunch of code that is
essentially unused beyond the basics of code shared between all types.

In the interest in keeping the code simple, migrate the few entries that are
using this measurement type to a generic measurement, then delete the tables
and code for "Phosphor". If we ever get back into seriously supporting the flow
cytometer, it can do the same thing with generic measurement types; and if
there's an external datasource / ontology we want to reference, we can revisit
adding back something that will explicitly reference that external source like
the other 'omics types.

Fixes: #1321