Fix MeasurementType admin creation page
The admin page to create new measurement type objects would display a
"Referenced in Studies" line for the not-yet-created measurement, and it would
list out any studies that did not yet have any measurements. Fix the function
returning the field list to display in the admin to exclude these studies on
the page for creating a new type entry.