Add nag email feature for import bulk create abuse
Hector was concerned about the checkboxes to bulk create lines or types in the
import, and their being abused by experimentalists who don't care about Hector
getting nice and clean data. Chris offered a suggestion to email whenever the
feature is used, and he promised to follow up with anyone who's using it for
evil for a short while. The code in this commit adds the email feature, and
tests for it. (And adds in a few more missing tests while I was in there).

Fixes: #1323