Fix some overlooked template items
As I was working on #1323, I found / re-discovered a potential gotcha in
templates, in dealing with the `blocktranslate` tags. While in most other parts
of templates using Django's templating language, items enclosed in double curly
brackets `{{}}` can have "filters" applied, text inside a `blocktranslate`
treats the text inside the brackets as a literal key to the template context.
The result will generally be a key not found in the context, and the variable
"expands" to an empty string. In the case of my re-discovery, this was in the
form of trying to do `{{ site_name|default:"EDD" }}`, so the site name would
print, and "EDD" would print if the name wasn't set. What really happens is
that instead of looking up a variable `site_name`, the template tries to look
up a variable `site_name|default:"EDD"` and yields the empty string (because
*of course* that "variable" doesn't exist).

While searching through templates for other potential cases of this error, I
also did a quick search on the `load` template tag, and where there were
multiples of these, consolidated into a single tag. For example `{% load i18n
%}{% load static %}` can be simply `{% load i18n static %}`.