Hotfix for Lines table display
When updating the TypeScript build to use the types shipped with the DataTables
library instead of the `@types` definitions, I accidentally removed the import
for the base library, as documentation showed importing only from the library
bindings for the UI library (Bootstrap 5 for us). But, this is breaking the
display of the Lines table on the Description view. Restoring the import fixes
the display.