Add custom Django admin site class
This is the first step to making a more flexible admin and settings page in
EDD, with views that are not directly tied to specific CRUD operations on
Django models in EDD. Specifically for tasks like in #290.

The basic implementation here adds very little changes to how the default admin
site operates, but does tweak a few minor details. The most obvious tweak is
replacing the Login / Logout views that come with the default admin site, and
instead using the same login as the rest of EDD. This has caused confusion with
the LBL Cybersecurity team in the past, which flagged the "different" admin
login page as a potential security issue. It also modifies some of the title
text to explicitly reference "EDD" rather than just a generic "Django".