Remove version attribute from compose files
Using `docker compose` with a recent Docker Desktop update of the sub-command,
is including a warning message about using the `version` top level attribute in
the compose file. This value is unused in Docker Compose, and the versions we
specified are actually referring to the versions for use with Docker Swarm and
`docker stack deploy`. It seems the maintainers are moving away from
version-based specifications, and are simply relying on the attributes in the
config file(s) matching the schema instead of a version number.

Since it's giving a warning now when launching with `docker compose up`, I'm
removing the attribute. We're using separate Swarm-specific files for running
in Jenkins and staging/production; the compose file here may need additional
tweaks if using `docker stack deploy` in development.