Update pre-commit hooks
Ran `pre-commit autoupdate` and then:

1. Added `ruff-format` rule in place of using `black`, since `ruff` has a
   drop-in replacement for `black`;
2. Removed the `black` pre-commit hook, since `ruff` now covers it;
3. Revert changes to `prettier` hook to keep the same version.

Note, the upstream repo for the `prettier` hook is now archived, the autoupdate
is trying to place in an alpha-version of `prettier`, and there seems to be a
bit of a flamewar over how to deal with `prettier`. It seems like something
broke compatibility, and the author of the hook got tired of responding to the
hordes that tried to autoupdate.