Remove GraphQL endpoint
We never really used the GraphQL endpoint, and the GraphiQL interface wasn't
really all that useful. Keeping it only serves to make a potentially larger
exploit surface, via a mechanism that isn't really tested. Also, the libraries
used have lagged behind Python and Django updates, and required shims over time
to keep the codebase compatible. It's best to just remove the feature now.

Also, as removing the GraphQL libraries requires re-locking dependencies,
pinning the version of Django Debug Toolbar to 4.3.0. The plugin updated with a
feature to try supporting a Django server running in Docker, but the
implementation is a little busted (detects Docker by assuming connections are
coming from the `.1` address of the host `/24`, and did the check by appending
the assumed address to the assumed list in settings `INTERNAL_IPS`--we use a
range object that implements `__contains__` instead of a list). There's a
patch, only the fix is not yet released.

Fixes: #1324