Upgrade to Django 5.0.x releases
Cleanup a handful of DeprecationWarnings that either already existed in EDD or
were triggered by an upgrade of Django to 5.0 from the previous bound of <4.2.

The Solr code was making use of `datetime.datetime.utcnow()`, which has been
deprecated for some time. The function returns a `datetime` with the current
UTC time, however, it returns it as a *naïve* `datetime`, without the `tzinfo`
timezone attribute set. This can cause issues, as Python defaults to treating
naïve `datetime` objects as in the system locale timezone. It's not impactful
in EDD's code, but regardless, the call is updated to use `arrow`'s functions
which will explicitly attach the UTC timezone to the object.

Some admin views would directly render an `Update` model to a string, and the
`__str__` method of the model uses the `arrow` library `humanize` method. This
would also make use of `datetime.datetime.utcnow()`, *unless* you explicitly
pass a time to compare. So, explicitly adding a non-deprecated comparison time.

The `edd.branding` module made use of Django's `django.utils.html.format_html`
function, which now warns if not passing an `args` or `kwargs` to fill
placeholders in the format string. EDD was doing this everywhere except for
default fallback values, using the empty string. These fallback values don't
need passing to `format_html`, can can simply return the empty string.

Django's default for URLs submitted in forms with `django.forms.URLField` is
changing from assuming an "http" scheme to assuming "https", and emits a
warning if a field is using the assumed scheme. There's a global setting for
this, added to EDD's base settings. The global setting itself is deprecated, as
it will be removed in Django 6.0; when upgrading to Django 6.0, the setting and
its warning filter should be removed.

See: https://blog.ganssle.io/articles/2019/11/utcnow.html
See: https://adamj.eu/tech/2023/12/07/django-fix-urlfield-assume-scheme-warnings/
Fixes: #1319