Replace FAQ entries for old Experiment Description
The new entry is for Experiment Setup, which recently replaced the Experiment
Description feature. The Experiment Setup files are more flexible versions of
the Experiment Description files, so example images of the files can remain.
Other images are updated for replacing "Experiment Description" text in the
image, or to use the newer, better-a11y, Bootstrap-5-based interface.

Update the dropzone area of the Study overview page to link to the FAQ entry
for the new Experiment Setup file.