Fix FAQ images to use Markdown images
The images in the FAQ page used hard-coded HTML IMG tags, which would result in
either the built documentation having broken image links, or the previews in
GitHub or GitLab having broken image links. Instead of doing that, use Markdown
syntax for the images, which should render correctly in the docs image,
`edd-docs.jbei.org`, Read The Docs, GitHub, GitLab, etc.

The images in the page were contained within HTML FIGURE tags, but they must be
moved out as the MkDocs parser will not convert the Markdown image into an
image tag if it's within literal HTML blocks.