Fix click handler conflict in docs link
When adding in the link to the EDD FAQ documentation for Experiment Setup, I
forgot to expand the click event interception to the link, as well as the
button. Clicking on the link would trigger the dropzone file selection dialog,
rather than going to the documentation. Moving the event handler up, and
explicitly including any A (link) HTML tags inside the `.dz-message` element
containing the user-readable instructions for the dropzone, and preventing the
click event from continuing to the dropzone click handler.