Replace drf-yasg with drf-spectacular
The API documentation generator we're using, `drf-yasg`, is not getting regular
updates. It points to `drf-spectacular` for new projects. Since we aren't doing
much special with the API docs, switching over now to get to use the OpenAPI 3
specifications. The generated pages under `drf-spectacular` seem to cover all
that was there before, and offers a bit more flexibility, and is updated.

Fixes: #1331