Refactor Experiment Setup data for assay splits
My initial implementation of Experiment Setup used a list of objects for all
metadata types, whether they were for Lines or Assays. These objects contained
a `protocol` field just in case the metadata was for an Assay, and needed a
Protocol lookup. Instead of keeping around a potentially empty field all the
time, instead remove `protocol` from the objects, and store metadata for Assays
in a `dict` that maps a Protocol UUID to a listing of metadata to use with that
protocol. This cleans up some messy bits of implementation, and makes the code
easier to follow.

By making this split, EDD can also now support creating multiple Assay records
per Line with Experiment Setup. The "simple" file-drop interface will only
support one Assay per Line/Protocol pairing; however, the planned REST API will
allow for advanced users to define lists of Assay metadata for multiple Assays.

Also in this refactoring are some small tweaks to handling of errors in saving,
so users see a red error progress bar instead of nothing in an error state.
Testing fixtures are cleaned up a bit, some old settings referenced but no
longer used are removed. The state transitions in the broker classes are moved
to the tasks instead, inside context managers. This simplifies the transition
code substantially and improves readability of broker mutation functions.