Fix coverage exclude for typing.Protocol methods
The formatter for `black` / `ruff` is now moving the empty `...` body for
methods on classes extending `typing.Protocol` to the same line as the `def`.
But, our config for `coverage` is only excluding lines from coverage if they
contain *only* spaces then `...`. So all our code using "interfaces" would get
marked as not covered. Removing the `^` from the exclude expression fixes, with
the minor side effect of also excluding any lines ending with a `# ...`
comment. At the moment, there are exactly two Python comments ending with `...`
and both are on lines with *only* a comment.

Be a force for good! Do not fudge coverage numbers by sprinkling `# ...` across
your code! If you truly need to exclude a line for some reason, be explicit and
use the built-in default of `# pragma: no cover`.