Update setup records to make name optional
The nested datastructures for Experiment Setup records include a "name" field,
to capture the metadata column heading or cell value for strains. To do the
REST API, it doesn't really make sense to refer to the "column heading", etc.
Some small tweaks allow for making these fields optional, so they don't need to
be needlessly included when going through the REST API.

Include additional tests to push coverage of the broker.py and parser.py files
to 100%.