Refactor redis cache interactions to utilities
In our refactorings of EDD's data processing code, we've moved to an
architecture of quick execution on HTTP responses, pushing almost all
computation into background workers with Celery, and mediating communication
between HTTP processing and background tasks with Redis cache keys. The
implementations of first data imports, then Experiment Description, each have
their own implementations of Redis interactions. These are similar
implementations, but with subtle differences as the code evolved to be better.

This refactor takes what we've learned from two separate implementations of
Redis communications in an upload / process / (un)resolved records workflow,
and pulls it out into a shared implementation of *only* the Redis primitive
operations. Both of the classes this is extracted from are trying to do lots of
(too many) things. Extracting this piece makes both classes much simpler, and
harmonizes implementations to contain the improvements we've learned about
since original authoring.

As a side-effect of the refactor, some issues with using Experiment Setup
failing to start the processing of uploaded files should be fixed.