Add tests for utility code
These changes add in tests for some of the utility classes and functions used
to operate EDD. Basically I was looking into the JSON serialization code for
REST API stuff, and got annoyed about the items in `edd/utilities.py` and
`edd/__init__.py` not having test coverage.

In the process, some of the settings are cleared up a bit, with the setting
responsible for determining absolute URLs getting pulled up into the
`__init__.py` of the settings module, where it detects a dev environment from
the `HTTPS_METHOD` environment used in Nginx proxying, and changes the settings
to use `http` instead of `https` when building absolute URLs. This also lets us
delete a hacky work-around function to do something similar with string
replacement at runtime.