Fix notifications in load/setup lock_status()
Part of refactoring the "request" classes in the `broker.py` modules of
`edd.load` and `edd.setup` was using a `lock_status()` function to handle doing
state transitions in tasks with a context manager. This simplified the code a
lot, but it also ended up dropping the final websocket updates that were
handled deep in the processing code. The fix here adds explicit updates after
the success/fail transitions in the `lock_status()` context manager.

Fixes: #1337