Fix regression in Compartment REST API
The Compartment API is a bit weird, because we don't actually have a database
model for them at this time; we use a hard-coded list of three possible values.
But, it's useful to have the front-end code query the server for the possible
values, for the time in the future where we may have more values (or a database
model + table). The prior way we were faking it had issues with generating a
REST API schema, so it updated to be a regular `ViewSet` instead of a faked
`ModelViewSet`. But, doing this removed the paging functionality that REST API
consumers in the data page are expecting. This updates the endpoint to include
a static "page" structure in the response.

Fixes: #1338