Update quickstart to use only one postgres
This brings the development setup script closer in line to what is currently
running in Jenkins. Instead of running one postgres container for EDD, and
another for ICE, put both databases in the same container.

Also ditching the SMTP container. We're just using `smtp.lbl.gov` internally,
and local development is using the console email handler. Since it's not going
to be maintained, dropping it rather than letting it sit and potentially break.