Add mock LDAP server to test auth code
For a long time, the only way to test out changes to our authentication code
would be to deploy, and test there. If it was production-focused authentication
code, that meant testing in production. But, I found a simple Dockerized LDAP
service, perfect for doing tests of LDAP authentication. That's allowed for
adding tests to fully cover the `edd.auth_backend` module (from 40%) and cover
a significant portion of the `edd.account` module.