Fix enforcement of Study default permissions
We were adding default permissions to Study records in the Django Form for
creating Studies. But, this skips over any other methods of Study creation
(e.g. the REST API). Moving the enforcement to the `save()` method of the model
makes sure the default permissions are set anywhere in the application.

While cleaning up the Study model, eliminated some dead code in the Comment,
Attachment, and Study models. There were a lot of method calls and properties
that were used in previous versions of the web interface, but are no longer
used in the current templates.