Move /users REST endpoint to /study/{id}/users
We were already requiring a study "fuzzy match" parameter to only list users on
a study where the requestor has access. Instead of keeping a top-level
endpoint, force selecting a specific Study and list out the users only in that
study. There's no need to allow enumerating users in the system.