Fix startup script error handling for Solr
During startup, if the Solr index checks fail, it was causing the entire
startup to abort. Wrapping this in a retry like the other connection checks
should allow it to complete once Solr is fully up.

The reindex tasks in startup appear to not function correctly when the startup
script runs the Django management command to reindex and then disowns the
process. So, instead of doing that, make the management command a fully
foreground process, but replace its guts with starting a Celery task to do the
reindex. Once the Celery workers are fully up, they will see the task in the
queue and kick of the reindex in the background.

Fixes: #1340