Fix email notification on account enable
When an admin user approves accounts, an email should go out to the user with
the approved account. If this is done individually on each profile page, it
works; if it's done from the list and the action dropdown, it was broken. The
action would do an `update()` on a queryset, then would loop the *same*
queryset to send emails; however, the `update()` and the loop both re-evaluate
the queryset, in a way where the loop would always be an empty queryset.
Forcing the queryset into a `list` before doing the `update` allows for
scheduling the email task for all the users affected.